Restructure patch
	- Split the project across multiple files.
	- Set up Cmake to build the project.
